split ResourcesRest into multiple files
